Search the Medicare Fee-for-Service Comprehensive Error Rate Testing API

Information on Medicare Fee-for-Service (FFS) claims that underwent Comprehensive Error Rate Testing (CERT) medical review. These claims were used to calculate the Medicare FFS improper payment rate.

Usage

codex_cert(
  year = 2021,
  part = NULL,
  drg = NULL,
  hcpcs = NULL,
  prov_type = NULL,
  bill_type = NULL,
  decision = NULL,
  error = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)

Source

Centers for Medicare & Medicaid Services

Arguments

year

YYYY, calendar year of CERT data. 2011-2021 data is currently available.

part

Type of Medicare Fee-for-Service claim

drg

The Diagnosis Related Group code

hcpcs

The Healthcare Common Procedure Coding System code

prov_type

Type of provider providing the service

bill_type

Type of Bill (TOB), Identifies type of facility, type of care, and sequence of bill in a particular episode of care; e.g. "721", "131", "110", "111"

decision

Medical review decision for the claim; either Agree or Disagree

error

Reason the claim was in error; e.g. "Insufficient Documentation", "Incorrect Coding"

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

Value

A tibble containing the search results.

Details

The Medicare Fee-for-Service (FFS) Comprehensive Error Rate Testing (CERT) dataset provides information on a random sample of FFS claims to determine if they were paid properly under Medicare coverage, coding, and payment rules. The dataset contains information on type of FFS claim, Diagnosis Related Group (DRG) and Healthcare Common Procedure Coding System (HCPCS) codes, provider type, type of bill, review decision, and error code. Please note, each reporting year contains claims submitted July 1 two years before the report through June 30 one year before the report. For example, the 2021 data contains claims submitted July 1, 2019 through June 30, 2020.

Links

Medicare Fee-for-Service Comprehensive Error Rate Testing

Note

Update Frequency: Annually

Examples

if (FALSE) {
codex_cert(hcpcs = "92002", decision = "Disagree")
}