Search the MedlinePlus Connect API

MedlinePlus Connect is a free service of the National Library of Medicine (NLM), National Institutes of Health (NIH), and the Department of Health and Human Services (HHS). This service allows health organizations and health IT providers to link patient portals and electronic health record (EHR) systems to MedlinePlus, an authoritative up-to-date health information resource for patients, families, and health care providers.

Usage

codex_medline(
  code = NULL,
  code_system = NULL,
  code_title = NULL,
  language = "english",
  format = "json"
)

Source

U.S. National Library of Medicine

Arguments

code

diagnosis (problem) code, medication code, laboratory test code, or procedure code

code_system

ICD-9-CM, ICD-10-CM, CPT, SNOMED, RXCUI, or NDC

code_title

You may also identify the name/title of the problem code. However, this information does not impact the response.

language

Identify if you would like the response to be in English or Spanish. MedlinePlus Connect will assume English is the language if it is not specified.

format

JSON, JSONP, or XML

Value

A tibble containing the search results.

Details

TMedlinePlus Connect accepts and responds to requests for information based on diagnosis (problem) codes, medication codes, laboratory test codes, and medical procedure codes. When an EHR, patient portal or other system submits a code-based request, MedlinePlus Connect returns a response that includes links to patient education information relevant to the code.

Note

MedlinePlus Connect API

Examples

codex_medline(code_title = "Chantix 0.5 MG Oral Tablet", code_system = "rxcui")
#> # A tibble: 1 × 7
#>   updated             entry_title source         searc…¹ resul…² entry…³ entry…⁴
#>   <dttm>              <chr>       <chr>          <chr>   <chr>   <chr>   <chr>  
#> 1 2023-01-05 17:37:11 Varenicline U.S. National… , RXNO… Medlin… https:… Vareni…
#> # … with abbreviated variable names ¹​search_description, ²​results_description,
#> #   ³​entry_link, ⁴​entry_summary

codex_medline(code = "00310-0751-39", code_system = "ndc")
#> # A tibble: 2 × 7
#>   updated             entry_title  source        searc…¹ resul…² entry…³ entry…⁴
#>   <dttm>              <chr>        <chr>         <chr>   <chr>   <chr>   <chr>  
#> 1 2023-01-05 17:37:11 Rosuvastatin U.S. Nationa… 00310-… Medlin… https:… "Rosuv…
#> 2 2023-01-05 17:37:11 Statins      U.S. Nationa… 00310-… Medlin… https:… "Stati…
#> # … with abbreviated variable names ¹​search_description, ²​results_description,
#> #   ³​entry_link, ⁴​entry_summary