Search CMS' NCCI Procedure to Procedure Edits (PTP) API

codex_ptp() allows you to search CMS' NCCI Procedure to Procedure Edits (PTP) API.

Usage

codex_ptp(
  category = NULL,
  column_1 = NULL,
  column_2 = NULL,
  modifier_indicator = NULL,
  ptp_edit_rationale = NULL,
  clean_names = TRUE,
  lowercase = TRUE,
  explain = FALSE
)

Source

Centers for Medicare & Medicaid Services

Arguments

category

Practitioner Services, Outpatient Hospital Services, DME Services

column_1

HCPCS comprehensive code

column_2

HCPCS component code

modifier_indicator

0 (Allowed), 1 (Not Allowed), 9 (Use Not Specified)

ptp_edit_rationale

Choices:

"Misuse of column two code with column one code"

"Standards of medical / surgical practice"

"CPT Manual or CMS manual coding instructions"

"Standard preparation / monitoring services for anesthesia"

"Mutually exclusive procedures"

"More extensive procedure"

"HCPCS/CPT procedure code definition"

"CPT \"separate procedure\" definition"

"Anesthesia service included in surgical procedure"

"Sequential procedure"

"Laboratory panel"

"Gender-specific (formerly Designation of sex) procedures"

clean_names

Clean column names with janitor's clean_names() function; default is TRUE.

lowercase

Convert column names to lowercase; default is TRUE.

explain

If TRUE, converts data to a more human-readable form; default is FALSE.

Value

A tibble containing the search results.

Details

The CMS National Correct Coding Initiative (NCCI) promotes national correct coding methodologies and reduces improper coding which may result in inappropriate payments of Medicare Part B claims and Medicaid claims. NCCI procedure-to-procedure (PTP) edits define pairs of Healthcare Common Procedure Coding System (HCPCS)/Current Procedural Terminology (CPT) codes that should not be reported together for a variety of reasons. The purpose of the PTP edits is to prevent improper payments when incorrect code combinations are reported. Practitioner services also refers to ambulatory surgical centers. DME refers to provider claims for durable medical equipment.

Links

CMS' NCCI Procedure to Procedure Edits (PTP) API

Note

Update Frequency: Quarterly

Examples

if (FALSE) {
codex_ptp(category = "DME Services")
codex_ptp(column_1 = "29000", modifier_indicator = "1")
codex_ptp(column_2 = "29015")
codex_ptp(modifier_indicator = "0")
codex_ptp(ptp_edit_rationale = "Mutually exclusive procedures")
}